first take on idt; there is a lot to be polished and understood too

+++ include/kernel/lib/bit.h
@@ -15,8 +15,8 @@ __force_inline bool test_bit(uint8_t bit_number, unsigned long long field) {
     return (field >> bit_number) & 1ULL;
 }
 
-__force_inline uint32_t extract_bit_chunk(uint8_t bit_from, uint8_t bit_to, uint32_t field) {
-    return (field >> bit_from) & ( UINT32_MAX >> (sizeof(uint32_t) * CHAR_BIT - bit_to - bit_from + 1));
+__force_inline uint64_t extract_bit_chunk(uint8_t bit_from, uint8_t bit_to, uint64_t field) {
+    return (field >> bit_from) & ( UINT64_MAX >> (sizeof(uint64_t) * CHAR_BIT - bit_to - bit_from + 1));
 }
 
 #endif /* INCLUDE_KERNEL_LIB_BIT_H_ */
